refactor: split theme constants into Shapes and FocusBorder
