+++ app/src/main/java/com/skiletro/wheelwitch/ui/components/FocusableSurface.kt
@@ -19,6 +19,7 @@ import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.skiletro.wheelwitch.ui.theme.WheelWitchPreviewTheme
+import com.skiletro.wheelwitch.ui.theme.buttonShape
 
 /**
  * A focusable card that draws a focus border when focused and optionally
