+++ app/src/main/java/com/skiletro/wheelwitch/ui/theme/Shapes.kt
@@ -1,0 +1,14 @@
+package com.skiletro.wheelwitch.ui.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.ui.unit.dp
+
+/**
+ * Shared corner radius used by primary action buttons throughout the app.
+ */
+val buttonShape = RoundedCornerShape(14.dp)
+
+/**
+ * Shared corner radius used by content sections (cards, panels, etc).
+ */
+val sectionShape = RoundedCornerShape(20.dp)
